revert 'calculate scrolltexty at compile-time' and change TextData.txtspd to int8_t 62734a61f410cc80f72cc2799a043b9fad7c58c8

+++ structs.h
@@ -482,7 +482,7 @@ typedef struct TextData {
 	const char* txtstr;
 	BOOLEAN scrlltxt;
 	BOOLEAN txtsfxset;
-	int txtspd;
+	int8_t txtspd;
 	int sfxnr;  // _sfx_id or sfx_set if txtsfxset is true
 } TextData;
